refactor: Fix most (true positive) Nicene errors
